+++ BirdsEye/sim/arduino_shim/avr/dtostrf.h
@@ -1,0 +1,4 @@
+#pragma once
+
+// dtostrf is declared (and implemented) by the Arduino shim.
+#include "../Arduino.h"